Refactor and simplify affix_stack.
The old implementation of affix_stack and affixstackinterface essentially
defined objects which from their signatures looked like Affix components.
I removed this level of indirection, and promoted the previous affix_misc
library (that does importing and builds stacks from strings) to be the
new affix_stack. It now returns the topmost (start-of-string) Affix
component, all configured and linked to the next component (also configured
and linked, etc.).

I've also streamlined what used to be affix_wrapper_lib's definition of
RepyNetworkApiWrapper, and refactored it into a new class of almost
identical name (but conforming to the Python naming standards,
RepyNetworkAPIWrapper). Much of the indirection has been removed, and
most object methods are now just references to the outer scope's
network calls. It remains to be seen if this Affix-like class is the
optimal implementation for what I'm trying to achieve (i.e. let the
bottommost Affix access the Repy network calls as transparently as
possible, without needing to check for the actual class name of the
next_affix).