Work in progress --- Made the AffixStack and RepyNetworkAPIWrapper use the Affix component interfaces.
The problem with my previous approach (let the topmost Affix component
represent the stack) had the problem that you couldn't push() new
components as new tops. Consequently, I re-added the stack manipulation
methods to the AffixStack. The stack also refreshes its network call
bindings when the top Affix component changes.

The RepyNetworkAPIWrapper now raises errors when you try to manipulate
the stack below it. (Since it sits below the bottom of the stack, there
is nothing you could manipulate there.)

Lastly, I cleaned up the BaseAffix a little so that it sports more
meaningful templates for the methods that derived classes need to
override. Further cleanup relates to the non-special-casing of the
RepyNetworkAPIWrapper in function calls. As a programmer, you will
except AffixStackError on pushes and pops, and you're good to go.