Make RepyNetworkAPIWrapper a subclass of BaseAffix; implement the methods defined by the Affix socket wrapper lib.
This is required for things like `socketobject.close()` which is
implemented as `socket_close(socketobject)` throughout the Affix
framework. (I'll explain in a docstring change in affix_wrapper_lib
that wrapping is needed for overriding methods of objects that are
yet to be instantiated). Since the wrapper lib's methods rely on
`peek()` to work, and RepyNetworkAPIWrapper doesn't allow that,
we need to override and call the actual Repy functions there.