Added resolving and name announcement to NamingAndResolverAffix.
* The network calls now try and resolve all [local,dest][host,ip]
arguments they get before calling the lower-layer Affix component.
* The constructor includes the logic for advertising a stack-id-to-ip-address
mapping (if a stack ID was given). My feeling is that this functionality
is useful enough so that it should live in a library, not an Affix component.
* I decided to make things not more complicated then necessary, so the
`resolve_identifier` function does not try to do recursive lookups and
detect loops. It can do key->IP address or list of addresses from the
advertise services, and plain gethostbyname DNS lookups. Pointing keys
to other keys to yet more keys and back again is not supported.