Made the NamingAndResolverAffix advertise its name.
My previous approach was that it should be a transparent Affix
component and that the remote side should not care whether it is
used on this stack or not. Also, the remote side could always decide
to use a NamingAndResolverAffix itself without anyone else's consent.

On the other hand, if one Affix stack uses it to advertise a stack ID,
then, in general, the other side is forced to use it too, no matter
if it learns that fact from the advertisement string, or by trying
without Resolver, then finding out it doesn't work, then trying
with Resolver.

So I guess the bottom line is that when you use NamingAndResolverAffix
to advertise a stack ID for you, it is wise to add it to the advertise
string as well.
(Alas, the absence of a stack ID in a NamingAndResolverAffix cannot
be taken to mean this Affix shouldn't place itself on the advertise
string: We instantiate copies of NamingAndResolverAffix on different
network calls such as `getconnetion`. By not setting their stack ID
on instantiation, we make them not advertise, and thus avoid spawning
lots of threads advertising the same stack-ID-to-IP-address mapping.)