Fix for SeattleTestbed/affix#57, and an additional minor fix.
The issue mentioned describes a fatal crash of `tcp_relay` when
a client drops a connection before the relay can send an error
message.

The second issue was with the inclusion of a `canilisten` test
in the relay startup code to check whether the desired port can
be contacted from the outside at all. The port number was wrongly
casted to `str`.