Skip Affix stack manipulations for designated Affix named ""
When MakeMeHearAffix has an empty `designated_affix_string`, it
must not try to register, push, listen on it, as the helper functions
in `affix_stack` don't allow empty Affix strings.

I believe checking and working around here (rather than in the lower
layers) is the sane thing to do, because the lower layers happen
to use the empty string as an indication that the bottom of the
stack has been reached; also, I don't see what the helper functions
should return at all if an Affix from an empty string is to be
instantiated in the middle of the stack.