Work around a possible KeyError in tcp_relay's registered_server dict.
It can happen that a server (i.e. host using a TCP relay to listen on
its behalf) is unregistered while a server connection request is being
handled, e.g. in line 340 of `tcp_relay.r2py`. With the dict entry
removed, the function processing the conn request previously lost
access to the `client_lock` stored in the server's dict entry,
resulting in the KeyError.

The bug behind this problem is that the unregister function uses
its own lock, not the server's `client_lock`, to serialize accesses
to the dict of registered servers. This needs further investigation.