Adding a basic LoggingAffix that logs function calls
This component was created partly for educational purposes, see the
GettingStartedWithAffix docs on the Seattle wiki where we use
LoggingAffix to explain a few concepts.

A few features are missing from this Affix at the moment, e.g.
tracking of LoggingAffix instances when new (partial) stacks are
instantiated / copied. At the moment, every LoggingAffix in every
stack will print "LoggingAffix: ...." (plus method call details),
making the ouput rather unclear when multiple LoggingAffixes are
employed at the same time.

The LoggingAffix howevere doesn't log the results of calls (or exceptions).
I think we'll want to have a separate DebugAffix for this.